refactor: 계산기 결과 값을 double -> int로 수정

+++ src/main/java/domain/Calculator.java
@@ -1,11 +1,11 @@
 package domain;
 
 public class Calculator {
-	public static double calculate(String[] values) {
-		double result = Double.parseDouble(values[0]);
+	public static int calculate(String[] values) {
+		int result = Integer.parseInt(values[0]);
 		for (int i = 1; i < values.length; i += 2) {
 			String operator = values[i];
-			double rightOperand = Double.parseDouble(values[i + 1]);
+			int rightOperand = Integer.parseInt(values[i + 1]);
 			result = Operator.of(operator)
 					.calculate(result, rightOperand);
 		}
